% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_gibbs.R
\name{coupled_gibbs}
\alias{coupled_gibbs}
\title{coupled_gibbs}
\usage{
coupled_gibbs(
  V,
  fieldfrequencies,
  hyper,
  algotuning,
  update.theta = FALSE,
  m = 1,
  lag = 1,
  max_iterations = Inf
)
}
\arguments{
\item{V}{a matrix representing the data}

\item{fieldfrequencies}{a list ...}

\item{hyper}{list containing hyperparameters}

\item{algotuning}{list containing precomputed and tuning parameters}

\item{m}{time horizon: the chains are sampled until the maximum between m and the meeting time; equal to one}

\item{lag}{time lag, equal to one by default}

\item{max_iterations}{maximum number of iterations, at which to interrup the while loop; Inf by default}
}
\description{
runs two chains, with a lag of at least one, and until either meeting occurs, or iteration 'm',
and at most 'max_iterations'
at the moment this only couples the update of eta and N
}
