% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_eta.R
\name{coupled_update_eta_relabel}
\alias{coupled_update_eta_relabel}
\title{coupled_update_eta_relabel}
\usage{
coupled_update_eta_relabel(state1, state2, V, algotuning)
}
\description{
performs max coupling of eta update as described in \code{\link{update_eta_relabel}}
}
