% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_theta.R
\name{coupled_update_theta}
\alias{coupled_update_theta}
\title{coupled_update_theta}
\usage{
coupled_update_theta(state1, state2, V, algotuning)
}
\value{
a list with 'state1' and 'state2'
}
\description{
coupled update of 'theta' component of two chains
using a mixture of random walk and independent MH, with common random numbers and maximal coupling ideas
}
