% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couplingdeduplication-package.R
\docType{package}
\name{couplingdeduplication-package}
\alias{couplingdeduplication-package}
\alias{couplingdeduplication}
\title{couplingdeduplication}
\description{
package to reproduce experiments in Tancredi et al's BA discussion paper 2020
}
\details{
...
}
\author{
...
}
\keyword{package}
