% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RLdata500.R
\name{get_RLdata500}
\alias{get_RLdata500}
\title{get_RLdata500}
\usage{
get_RLdata500()
}
\value{
a list with 'V', 'Mvec', and 'fieldfrequencies'
}
\description{
loads the RecordLinkage and the 'RLdata500' data set
which contains 500 rows and 14 fields
The data is the matrix V, the vector 'Mvec'
contains the number of possible categories for each of the 14 fields
and 'fieldfrequencies' is a list where each of the 14 elements
is a vector of frequencies of occurrence of these different possible categories.
}
