% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcoupling_normals.R
\name{reflectionmaxcoupling}
\alias{reflectionmaxcoupling}
\title{Reflection-maximal coupling of two univariate Normal distributions}
\usage{
reflectionmaxcoupling(mu1, mu2, sigma)
}
\arguments{
\item{mu1}{First mean}

\item{mu2}{Second mean}

\item{sigma}{Common standard deviation}
}
\value{
Returns a list with
\itemize{

\item "xy": the pair of samples \eqn{(x,y)}

\item "identical": TRUE if \eqn{x = y}, FALSE otherwise
}
}
\description{
Sample from reflection-maximal coupling of two univariate Normal distributions,
specified through their means, with common standard deviation.
See \code{\link{rmvnorm_reflectionmax}} for a multivariate version.
}
