% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel}
\alias{relabel}
\title{relabel}
\usage{
relabel(eta, clustering)
}
\arguments{
\item{eta}{vector of n labels}

\item{clustering}{list representing a partition, with members 'clmembers', 'clsize' as e.g. create by \code{init_clustering}}
}
\value{
a list with 'eta' and 'old_to_new', which is a vector that can be used to permute vectors or matrices indexed by labels
}
\description{
function to perform re-labeling;
goes through each cluster and gives it the label
of smallest index among members of the cluster
}
