% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinit.R
\name{rinit}
\alias{rinit}
\title{rinit}
\usage{
rinit(n, fieldfrequencies, hyper, V)
}
\arguments{
\item{n}{the number of rows in the data}

\item{fieldfrequencies}{frequencies of appearance of each element in each field (list of vectors), 
such as provided by \code{\link{get_RLdata500}}}

\item{hyper}{a list of hyperparameters including m0, s0, s
where m0 is the mean of beta0, s0 its standard deviation, s is the std dev of beta_diff}
}
\description{
initialization of the chains...
}
