% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_gibbs.R
\name{single_gibbs}
\alias{single_gibbs}
\title{single_gibbs}
\usage{
single_gibbs(
  nmcmc,
  V,
  fieldfrequencies,
  hyper,
  algotuning,
  update.beta = FALSE,
  update.theta = FALSE
)
}
\value{
a list with  'ksize_history', 'N_history' and 'theta_history'
}
\description{
runs a Gibbs sampler with updates of eta, N, and optionally of theta
}
