% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{update_beta}
\alias{update_beta}
\title{update_beta}
\usage{
update_beta(state, hyper, V, algotuning)
}
\value{
a state with updated 'beta_0', 'beta_diff', 'alpha' and updated loglikelihood
}
\description{
update 'beta' component of the chain, that is both beta_0 and beta_diff
}
