% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_eta.R
\name{update_eta_relabel}
\alias{update_eta_relabel}
\title{update_eta_relabel}
\usage{
update_eta_relabel(state, V, algotuning)
}
\arguments{
\item{state}{a list representing the state of the Markov chain as e.g. generate by \code{\link{rinit}}}

\item{V}{a matrix representing the data}

\item{algotuning}{a list containing 'eta_update_prob'}
}
\value{
state of the Markov chain with new eta and partition
}
\description{
update eta given the rest, and perform "relabeling" of the partition
given the other components of the chain. Update each eta with some probability as indicated by 
'eta_update_prob' in 'algotuning'
}
