% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_N.R
\name{update_N}
\alias{update_N}
\title{update_N}
\usage{
update_N(state, V, hyper, algotuning)
}
\arguments{
\item{state}{a list representing the state of the Markov chain as e.g. generate by \code{\link{rinit}}}

\item{V}{a matrix representing the data}

\item{hyper}{a list containing hyperparameter 'g'}

\item{algotuning}{a list containing pre-computed quantities 'N_max', 'lfactorials', 'lns'}
}
\value{
state of the Markov chain with updated 'N'
}
\description{
takes a state of the Markov chain and draws a new N.
The procedure truncates the distribution to 'N_max', computes the probabilities, and uses the 'sample.int' function of R
}
