% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_theta.R
\name{update_theta}
\alias{update_theta}
\title{update_theta}
\usage{
update_theta(state, V, algotuning)
}
\value{
a state with updated 'theta' and updated loglikelihood
}
\description{
update 'theta' component of the chain
using a mixture of random walk and independent MH
}
