% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couplingdeduplication-package.R
\docType{package}
\name{couplingdeduplication-package}
\alias{couplingdeduplication-package}
\alias{couplingdeduplication}
\title{couplingdeduplication}
\description{
package to reproduce experiments in Tancredi, Steorts, Liseo BA discussion paper 2020
"A Unified Framework for De-Duplication and Population Size Estimation"
and to implement coupling of main Gibbs sampler... URL https://github.com/EmiliaPompe/discussion_unified_framework
}
\details{
...
}
\author{
...
}
\keyword{package}
